revised build process so that the skins folder is copied to the build directory and loaded at runtime, instead of being compiled to a resource file
git-svn-id: http://svn.powertabs.org/svn/ptedev/trunk@37 5b1460fe-822e-4aa1-ac1b-8a6412c92c8c

+++ source/documentmanager.cpp
@@ -21,10 +21,10 @@ DocumentManager::~DocumentManager()
 
 PowerTabDocument* DocumentManager::getCurrentDocument()
 {
-	if (currentDocumentIndex < 0 || currentDocumentIndex >= documentList.size())
-		return NULL;
-	else
-		return documentList[currentDocumentIndex];
+    if (currentDocumentIndex < 0 || (int) currentDocumentIndex >= documentList.size())
+        return NULL;
+    else
+        return documentList[currentDocumentIndex];
 }
 
 void DocumentManager::Remove(uint32_t index)
